backend/pyproject.toml
[project]
name = "comp-intel-backend"
version = "0.1.0"
description = "FastAPI backend for Competitive Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "supabase>=2.0.0",
    "tavily-python>=0.5.0",
    "firecrawl-py>=0.0.16",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "schemas", "agent"]
